Handle missing config keys during checkpoint search (#10)
* Handle missing config keys during checkpoint search

When queried for checkpoints of a root graph, the `config` object the
checkpointer receives will not have a `checkpoint_id` or `checkpoint_ns`
key. Currently, we interpret their absence incorrectly by searching for the 
string literal "None", which doesn't find a matching checkpoint.

Unfortunately, older RediSearch versions don't support the INDEXEMPTY
and INDEXMISSING options introduced in version 2.10, and the library this
package uses, RedisVL, doesn't support those options yet.

In this PR, we introduce sentinel values for strings and IDs that allow
querying for empty values. This solves the problem with some loss of
elegance and grace.